backend/requirements.txt
fastapi~=0.115.6
pydantic~=2.10.4
sqlmodel~=0.0.22
pydantic-settings~=2.7.1
aiosqlite~=0.20.0
pyjwt==2.10.1
SQLAlchemy~=2.0.36
pytest~=8.3.4
passlib~=1.7.4